+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- CurrentWidget::find_block() to resolve a block widget's named block
+
 ## [1.2.0] - 2026-07-09
 
 ### Added